# lplates/openapi.yaml
openapi: 3.0.3
info:
  title: lplates API
  description: >
    REST API for the lplates driving-lesson booking platform. Covers user
    authentication, instructor management, lesson bookings, EDT (Essential
    Driver Training) progress tracking, instructor pricing, and feedback.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get the current authenticated user's basic info
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/profile:
    get:
      tags: [Users]
      summary: Get the current user's role-specific profile
      description: >
        Returns `LearnerProfile` for LEARNER, `InstructorProfile` for
        INSTRUCTOR, or a plain `User` for ADMIN.
      responses:
        '200':
          description: Role-specific profile
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LearnerProfile'
                  - $ref: '#/components/schemas/InstructorProfile'
                  - $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/:
    get:
      tags: [Users]
      summary: Get all users (ADMIN only)
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get basic user info by ID (ADMIN only)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/profile:
    get:
      tags: [Users]
      summary: Get role-specific profile for any user (ADMIN only)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Role-specific profile
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LearnerProfile'
                  - $ref: '#/components/schemas/InstructorProfile'
                  - $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid username or password"
    Role:
      type: string
      enum: [LEARNER, INSTRUCTOR, ADMIN]
    LessonType:
      type: string
      enum: [lesson, edt, test_car_hire]
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected]
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Instructor:
      type: object
      properties:
        instructorId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        approvalStatus:
          $ref: '#/components/schemas/ApprovalStatus'
        gender:
          type: string
        phoneNumber:
          type: string
        adiNumber:
          type: string
        transmission:
          type: string
          example: manual
        yearsExperience:
          type: integer
        rating:
          type: number
          format: double
        carMake:
          type: string
        carModel:
          type: string
        locations:
          type: array
          items:
            type: string
        profilePicture:
          type: string
          nullable: true
          description: URL relative to the API base, e.g. /uploads/instructors/1/abc.jpg
        description:
          type: string
        reviewsCount:
          type: integer
        agreeTerms:
          type: boolean
        offersTestCarHire:
          type: boolean
        testCarHirePrice:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
...
    Lesson:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        scheduledDate:
          type: string
          format: date
          example: "2026-08-10"
        scheduledTime:
          type: string
          example: "09:00:00"
        durationMinutes:
          type: integer
          example: 60
        status:
          type: string
          example: pending
        paymentStatus:
          type: string
          example: unpaid
        price:
          type: number
          format: double
          example: 60.00
        currency:
          type: string
          example: EUR
        lessonType:
          $ref: '#/components/schemas/LessonType'
        edtModule:
          type: string
        edtCompleted:
          type: boolean
        notes:
          type: string
        cancellationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EdtProgress:
      type: object
      properties:
        id:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        moduleNumber:
          type: integer
          minimum: 1
          maximum: 12
        moduleName:
          type: string
        completed:
          type: boolean
        lessonId:
          type: integer
          format: int64
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    EdtSummary:
      type: object
      properties:
        totalModules:
          type: integer
          example: 12
        completedModules:
          type: integer
        remainingModules:
          type: integer
        fullyCompleted:
          type: boolean
    LearnerProfile:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
        edtProgress:
          type: array
          items:
            $ref: '#/components/schemas/EdtProgress'
        edtSummary:
          $ref: '#/components/schemas/EdtSummary'
    InstructorProfile:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        instructor:
          $ref: '#/components/schemas/Instructor'